requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.7.0  # For LoRA fine-tuning
bitsandbytes>=0.41.0  # For 8-bit training
scikit-learn>=1.3.0  # For dataset splitting

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
wandb>=0.15.0  # For experiment tracking (optional)

# Development
black>=23.0.0  # Code formatting
pytest>=7.0.0  # Testing 